+++ graph_utils/graph_generator_test.cc
@@ -56,23 +56,23 @@ private:
 
 TEST_F(SimpleGraphGeneratorTest, DegreeSequence) {
   {
-    vector<int> v({ 1, 1 });
+    vector<int> v({1, 1});
     EXPECT_TRUE(SimpleGraphGenerator::IsGraphicalDegreeSeq(v));
   }
   {
-    vector<int> v({ 3, 3, 2, 2, 2, 2, 2, 2 });
+    vector<int> v({3, 3, 2, 2, 2, 2, 2, 2});
     EXPECT_TRUE(SimpleGraphGenerator::IsGraphicalDegreeSeq(v));
   }
   {
-    vector<int> v({ 1, 1, 1 });
+    vector<int> v({1, 1, 1});
     EXPECT_FALSE(SimpleGraphGenerator::IsGraphicalDegreeSeq(v));
   }
   {
-    vector<int> v({ 4, 4, 2, 1, 1 });
+    vector<int> v({4, 4, 2, 1, 1});
     EXPECT_FALSE(SimpleGraphGenerator::IsGraphicalDegreeSeq(v));
   }
   {
-    vector<int> v({ 4, 4, 5, 5, 2 });
+    vector<int> v({4, 4, 5, 5, 2});
     // The degree sequence should be in non-increasing order.
     ASSERT_THROW(SimpleGraphGenerator::IsGraphicalDegreeSeq(v),
                  std::invalid_argument);
@@ -81,30 +81,30 @@ TEST_F(SimpleGraphGeneratorTest, DegreeSequence) {
 
 TEST_F(SimpleGraphGeneratorTest, ReducingDegreeSequence) {
   {
-    vector<int> seq({ 1, 1 });
+    vector<int> seq({1, 1});
     vector<int> incident;
     // No vertices are incident on 1, but it has a degree of 1.
     ASSERT_THROW(SimpleGraphGenerator::ReduceDegreeSequence(0, incident, &seq),
                  std::invalid_argument);
   }
   {
-    vector<int> seq({ 1, 1 });
-    vector<int> incident({ 2 });
+    vector<int> seq({1, 1});
+    vector<int> incident({2});
     // Invalid vertex 2.
     ASSERT_THROW(SimpleGraphGenerator::ReduceDegreeSequence(0, incident, &seq),
                  std::invalid_argument);
   }
   {
-    vector<int> seq({ 1, 1 });
-    vector<int> incident({ 1 }); // 1 is incident on 0.
+    vector<int> seq({1, 1});
+    vector<int> incident({1}); // 1 is incident on 0.
     SimpleGraphGenerator::ReduceDegreeSequence(0, incident, &seq);
     ASSERT_EQ(2, seq.size());
     EXPECT_EQ(0, seq[0]); // This vertex has been removed.
     EXPECT_EQ(0, seq[1]); // This vertex was incident on the removed vertex.
   }
   {
-    vector<int> seq({ 4, 4, 2, 1, 1 }); // Non-graphical sequence is fine.
-    vector<int> incident({ 0, 2, 3, 4 });
+    vector<int> seq({4, 4, 2, 1, 1}); // Non-graphical sequence is fine.
+    vector<int> incident({0, 2, 3, 4});
     SimpleGraphGenerator::ReduceDegreeSequence(1, incident, &seq);
     ASSERT_EQ(5, seq.size());
     EXPECT_TRUE(seq[1] == 0 && seq[3] == 0 && seq[4] == 0);
@@ -115,15 +115,15 @@ TEST_F(SimpleGraphGeneratorTest, ReducingDegreeSequence) {
 
 TEST_F(SimpleGraphGeneratorTest, ConstrainedGraphicalityTest) {
   {
-    vector<pair<int, bool> > seq({ make_pair(3, true), make_pair(3, false),
-                                   make_pair(2, true), make_pair(2, true),
-                                   make_pair(2, true), make_pair(2, true),
-                                   make_pair(2, true), make_pair(2, true) });
+    vector<pair<int, bool>> seq({make_pair(3, true), make_pair(3, false),
+                                 make_pair(2, true), make_pair(2, true),
+                                 make_pair(2, true), make_pair(2, true),
+                                 make_pair(2, true), make_pair(2, true)});
     EXPECT_TRUE(SimpleGraphGenerator::CGTest(0, seq));
   }
   {
-    vector<pair<int, bool> > seq({ make_pair(2, true), make_pair(2, true),
-                                   make_pair(1, true), make_pair(1, true) });
+    vector<pair<int, bool>> seq({make_pair(2, true), make_pair(2, true),
+                                 make_pair(1, true), make_pair(1, true)});
     EXPECT_TRUE(SimpleGraphGenerator::CGTest(0, seq));
     seq[2].second = false;
     EXPECT_TRUE(SimpleGraphGenerator::CGTest(0, seq));
@@ -133,9 +133,9 @@ TEST_F(SimpleGraphGeneratorTest, ConstrainedGraphicalityTest) {
 }
 
 TEST_F(SimpleGraphGeneratorTest, GenerateAllAdjSetsTest) {
-  vector<int> seq({ 3, 3, 2, 2, 2, 2, 2, 2 });
+  vector<int> seq({3, 3, 2, 2, 2, 2, 2, 2});
   set<int> cur_set;
-  vector<set<int> > all_adj_sets;
+  vector<set<int>> all_adj_sets;
   SimpleGraphGenerator::GenerateAllAdjSets(seq, &cur_set, &all_adj_sets);
 
   ASSERT_EQ(35, all_adj_sets.size());
@@ -145,7 +145,7 @@ TEST_F(SimpleGraphGeneratorTest, GenerateAllAdjSetsTest) {
       for (int k = j - 1; k > 0; k--) {
         const set<int> &cur_set = all_adj_sets[entry++];
         int counter = 0;
-        int expected[3] = { k, j, i };
+        int expected[3] = {k, j, i};
         for (auto it = cur_set.cbegin(); it != cur_set.cend(); it++) {
           EXPECT_EQ(expected[counter++], *it);
         }
@@ -153,9 +153,9 @@ TEST_F(SimpleGraphGeneratorTest, GenerateAllAdjSetsTest) {
 }
 
 TEST_F(SimpleGraphGeneratorTest, GenerateAllAdjSetsTest2) {
-  vector<int> seq({ 2, 2, 1, 1 });
+  vector<int> seq({2, 2, 1, 1});
   set<int> cur_set;
-  vector<set<int> > all_adj_sets;
+  vector<set<int>> all_adj_sets;
   SimpleGraphGenerator::GenerateAllAdjSets(seq, &cur_set, &all_adj_sets);
 
   ASSERT_EQ(2, all_adj_sets.size());
@@ -175,12 +175,12 @@ TEST_F(SimpleGraphGeneratorTest, GenerateAllAdjSetsTest2) {
 }
 
 TEST_F(SimpleGraphGeneratorTest, GenerateAllGraphsByDegreeSeq) {
-  vector<int> seq({ 2, 2, 1, 1 });
+  vector<int> seq({2, 2, 1, 1});
   vector<Graph *> v;
   SimpleGraphGenerator::GenerateAllGraphs(seq, &v);
   ASSERT_EQ(2, v.size());
-  vector<string> expected1({ "0101", "1010", "0100", "1000" });
-  vector<string> expected2({ "0110", "1001", "1000", "0100" });
+  vector<string> expected1({"0101", "1010", "0100", "1000"});
+  vector<string> expected2({"0110", "1001", "1000", "0100"});
   vector<string> result1;
   vector<string> result2;
   v[0]->GetAdjMatrix(&result1);
@@ -191,18 +191,18 @@ TEST_F(SimpleGraphGeneratorTest, GenerateAllGraphsByDegreeSeq) {
 
 TEST_F(SimpleGraphGeneratorTest, GenerateAllUniqueGraphsByDegreeSeq) {
   {
-    vector<int> seq({ 2, 2, 1, 1 });
+    vector<int> seq({2, 2, 1, 1});
     vector<Graph *> v;
     DummyFilter filter(true); // All graphs are allowed to pass.
     SimpleGraphGenerator::GenerateAllUniqueGraphs(seq, &filter, &v);
     ASSERT_EQ(1, v.size());
-    vector<string> expected({ "0101", "1010", "0100", "1000" });
+    vector<string> expected({"0101", "1010", "0100", "1000"});
     vector<string> result;
     v[0]->GetAdjMatrix(&result);
     ExpectVectorsEq<string>(expected, result);
   }
   {
-    vector<int> seq({ 2, 2, 1, 1 });
+    vector<int> seq({2, 2, 1, 1});
     vector<Graph *> v;
     DummyFilter filter(false); // No graph is allowed to pass.
     SimpleGraphGenerator::GenerateAllUniqueGraphs(seq, &filter, &v);
@@ -212,33 +212,32 @@ TEST_F(SimpleGraphGeneratorTest, GenerateAllUniqueGraphsByDegreeSeq) {
 
 TEST_F(SimpleGraphGeneratorTest, GenerateAllDegreeSeqs) {
   {
-    vector<vector<int> > seqs;
+    vector<vector<int>> seqs;
     SimpleGraphGenerator::GenerateAllDegreeSequences(3, &seqs);
-    vector<vector<int> > expected(
-        { { 2, 2, 2 }, { 2, 2, 1 }, { 2, 1, 1 }, { 1, 1, 1 } });
-    for (int i = 0; i < seqs.size(); ++i) {
+    vector<vector<int>> expected({{2, 2, 2}, {2, 2, 1}, {2, 1, 1}, {1, 1, 1}});
+    for (size_t i = 0; i < seqs.size(); ++i) {
       ExpectVectorsEq<int>(expected[i], seqs[i]);
     }
   }
   {
-    vector<vector<int> > seqs;
+    vector<vector<int>> seqs;
     SimpleGraphGenerator::GenerateAllDegreeSequences(4, &seqs);
-    vector<vector<int> > expected({ { 3, 3, 3, 3 },
-                                    { 3, 3, 3, 2 },
-                                    { 3, 3, 3, 1 },
-                                    { 3, 3, 2, 2 },
-                                    { 3, 3, 2, 1 },
-                                    { 3, 3, 1, 1 },
-                                    { 3, 2, 2, 2 },
-                                    { 3, 2, 2, 1 },
-                                    { 3, 2, 1, 1 },
-                                    { 3, 1, 1, 1 },
-                                    { 2, 2, 2, 2 },
-                                    { 2, 2, 2, 1 },
-                                    { 2, 2, 1, 1 },
-                                    { 2, 1, 1, 1 },
-                                    { 1, 1, 1, 1 } });
-    for (int i = 0; i < seqs.size(); ++i) {
+    vector<vector<int>> expected({{3, 3, 3, 3},
+                                  {3, 3, 3, 2},
+                                  {3, 3, 3, 1},
+                                  {3, 3, 2, 2},
+                                  {3, 3, 2, 1},
+                                  {3, 3, 1, 1},
+                                  {3, 2, 2, 2},
+                                  {3, 2, 2, 1},
+                                  {3, 2, 1, 1},
+                                  {3, 1, 1, 1},
+                                  {2, 2, 2, 2},
+                                  {2, 2, 2, 1},
+                                  {2, 2, 1, 1},
+                                  {2, 1, 1, 1},
+                                  {1, 1, 1, 1}});
+    for (size_t i = 0; i < seqs.size(); ++i) {
       ExpectVectorsEq<int>(expected[i], seqs[i]);
     }
   }
